Fix case mangling in ManifestParser
ManifestParser converts text to lowercase when matching directives,
which causes text associated with them (ex: the inside of the parens
for an if statement) to be passed as lowercase as well.  This can cause
conditional evaluation to fail because the case of identifiers is
changed.

To fix, remove the lowercase conversion (there is no need to allow
anycase directive matching in manifest files; it is not allowed in
other files, either.

Type: Bug Fix
Signed-off-by: Daniel Kristensen <daniel.kristensen@garmin.com>